+++ CHANGELOG.md
@@ -3,3 +3,7 @@
 ## Version 120000.01
 
 - The initial release.
+- Ported project from LS: Glass
+- Updated the code to World of Warcraft: Midnight
+- Introduced timestamp option
+- Introduced "Debug" mode for the addon